atarist: rework lookup of keys
The previous logic for mapping keys did not work when
using eg. an us TOS version with a german keyboard:

- when pressing "-" (scancode 0x35 on german keyboard), that
got mapped to scancode 0x12, which in turn caused the us version
to map it to "-" again.
- but when typing ":" (shifted scancode 0x34 above "."), that sends
0x34 to the TOS (they happen to have the same location), which then
generates ">" (the mapping for an us keyboard).

That results in an ugly and confusing mix of german/us assignments.
Even worse, that results in some characters (eg. "/") not being
available at all ("/" is in the location of that "-" on an us keyboard,
and typing shift-7 on a german keyboard results in "&")

That was changed now to use just the scancodes, which result in an us
keyboard layout when using an us TOS version.